agent/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# LangChain & LangGraph
# IMPORTANT: These versions are tested and compatible together
# DO NOT upgrade individually - upgrade all together if needed
langchain==0.1.20
langchain-core>=0.1.52,<0.2.0  # Will install latest 0.1.x (0.1.53)
langchain-community==0.0.38
langchain-openai==0.1.7
langchain-anthropic==0.1.11
langchain-text-splitters==0.0.1
langchain-qdrant==0.1.0
langgraph==0.0.20  # Compatible with langchain-core 0.1.x
langsmith>=0.1.55,<0.2.0  # Will install latest 0.1.x (0.1.147)

# Vector DB & Embeddings
qdrant-client==1.9.1
openai>=1.24.0,<2.0.0
anthropic>=0.23.0,<1.0.0
tiktoken>=0.7.0,<1.0.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0
numpy>=1.23.2,<2.0.0  # Required for pandas compatibility
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.2.0
python-magic==0.4.27  # Works with libmagic1 installed via Dockerfile
# python-magic-bin==0.4.14

# OCR
pytesseract==0.3.10

# Tools
duckduckgo-search==4.1.1
wikipedia==1.4.0

# Utils
httpx==0.26.0
aiohttp==3.9.1

# Redis Cache
redis==5.0.1

# Retry Logic
tenacity==8.2.3
